# openapi/kayan_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kayan Orphan Care — AI Agent Platform (Mock)",
    "description": "Prototype backend simulating جمعية كيان للايتام beneficiary management. Multi-channel (Voice/SIP + WhatsApp), Arabic-first. Each route is an agent tool.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/crm/tickets": {
      "get": {
        "tags": [
          "2 · نظام التذاكر CRM | Tickets & Kanban"
        ],
        "summary": "List tickets with filters",
        "description": "Lists tickets (التذاكر والطلبات). Filter by status, department, channel or beneficiary. Each row carries the SLA countdown like the admin panel.",
        "operationId": "list_tickets_crm_tickets_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "open"
              ],
              "title": "Status"
            }
          },
          {
            "name": "department_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "DEP-BEN"
              ],
              "title": "Department Id"
            }
          },
          {
            "name": "channel",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "examples": [
                "whatsapp"
              ],
              "title": "Channel"
            }
          },
          {
            "name": "beneficiary_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Beneficiary Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "2 · نظام التذاكر CRM | Tickets & Kanban"
        ],
        "summary": "Open a ticket from any channel",
        "description": "Creates a ticket from a WhatsApp message or a phone call. If a phone is given the beneficiary is auto-matched. This is what the bot calls when it cannot resolve the query itself or when a human must act.",
        "operationId": "create_ticket_crm_tickets_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crm/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "2 · نظام التذاكر CRM | Tickets & Kanban"
        ],
        "summary": "Get a ticket with its full conversation",
        "description": "Full ticket detail including the message log (سجل المحادثة), SLA remaining, and the beneficiary's other tickets — the same context the agent screen shows.",
        "operationId": "get_ticket_crm_tickets__ticket_id__get",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crm/tickets/{ticket_id}/status": {
      "patch": {
        "tags": [
          "2 · نظام التذاكر CRM | Tickets & Kanban"
        ],
        "summary": "Move a ticket across the kanban board",
        "description": "Moves a ticket to another column (مفتوح / جاري العمل / بانتظار العميل / تم الرد / مغلق).",
        "operationId": "move_ticket_crm_tickets__ticket_id__status_patch",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveTicketIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crm/tickets/{ticket_id}/assign": {
      "patch": {
        "tags": [
          "2 · نظام التذاكر CRM | Tickets & Kanban"
        ],
        "summary": "Assign a ticket to a staff member",
        "description": "Assigns the ticket to a member of فريق العمل.",
        "operationId": "assign_ticket_crm_tickets__ticket_id__assign_patch",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/crm/tickets/{ticket_id}/reply": {
      "post": {
        "tags": [
          "2 · نظام التذاكر CRM | Tickets & Kanban"
        ],
        "summary": "Post a reply on a ticket",
        "description": "Adds an outbound message to the conversation and moves the ticket to 'replied'. If the WhatsApp 24h window has expired the response warns that only a template message may be sent.",
        "operationId": "reply_ticket_crm_tickets__ticket_id__reply_post",
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignIn": {
        "properties": {
          "staff_id": {
            "type": "string",
            "title": "Staff Id",
            "examples": [
              "STF-02"
            ]
          }
        },
        "type": "object",
        "required": [
          "staff_id"
        ],
        "title": "AssignIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MoveTicketIn": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "open | in_progress | waiting_customer | replied | closed | expired",
            "examples": [
              "in_progress"
            ]
          },
          "note_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note Ar"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "MoveTicketIn"
      },
...
      "ReplyIn": {
        "properties": {
          "body_ar": {
            "type": "string",
            "title": "Body Ar",
            "examples": [
              "تم استلام طلبكم وجاري دراسته"
            ]
          },
          "sender": {
            "type": "string",
            "title": "Sender",
            "default": "agent",
            "examples": [
              "agent",
              "bot"
            ]
          }
        },
        "type": "object",
        "required": [
          "body_ar"
        ],
        "title": "ReplyIn"
      },
...
    }
  }
}